Starting to Separate the "Player" from the "Connection"
PlayerAddress was changed to be a string so we do not have to keep parsing and .ToString everywhere we want to use it.  (SignalR connectionID is a guid saved as a string...)

When Creating a game, we must specify if this is going to be a muliplayer or single player game.